modifica classi
