+++ Extranet/src/app/classi/ClassiUtenti.ts
@@ -1,0 +1,19 @@
+export class Utente
+{
+  constructor(
+  public id ?: number ,
+  public username  ?: string,
+  public password ?:  string,
+  public stato ?: number,
+  public primo_accesso ?: number,
+  public gruppo ?: Gruppo
+  ){}
+}
+
+export class Gruppo
+{
+  constructor(
+   public id  ?: number,
+   public descrizione ?: string
+    ) {}
+}
