+++ Extranet/src/app/dati/dati.component.ts
@@ -1,8 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
-import { Risorse } from '../assegnarisorse/assegnarisorse.component';
-import { Risorsa } from '../preventivorisorse/preventivorisorse.component';
+import { Risorsa, Risorse } from '../classi/ClassiGenerali';
 import { InserimentoService } from '../services/inserimento.service';
 
 @Component({
