+++ Extranet/src/app/modificaavanzamento/modificaavanzamento.component.ts
@@ -1,14 +1,11 @@
 import { ThisReceiver } from '@angular/compiler';
 import { Component, OnInit , ViewChild } from '@angular/core';
-import { Avanzamento, TipoAvanzamento } from '../inseriscitask/inseriscitask.component';
 import { InserimentoService } from '../services/inserimento.service';
-import { Mese,Anno } from '../inseriscitask/inseriscitask.component';
-import { Attivita } from '../assegnatask/assegnatask.component';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatPaginator } from '@angular/material/paginator';
 import { ActivatedRoute } from '@angular/router';
-import { Commessa } from '../selezionacommessa/selezionacommessa.component';
 import { SessionStorageService } from 'angular-web-storage';
+import { Anno, Attivita, Avanzamento, Commessa, Mese, TipoAvanzamento } from '../classi/ClassiGenerali';
 
 
 @Component({
